app/requirements.txt
streamlit==1.30.0
tensorflow==2.15.0.post1
numpy==1.26.3
Pillow
pandas
plotly
requests
supabase
